+++ example-crs-architecture/CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG.md
 
+## 2025-03-07
+
+- Adds tailscale configuration for ingress
+- Removes nginx-ingress component
+- Removes cert-manager / letsencrypt component
+- Removes cluster issuer component
+- Updates to env variables
+- Updates to README.md
+
 ## 2025-03-01
 
 - Modified variable names